apps/backend/models/lactation-autoencoder/pyproject.toml
[project]
name = "bovi-lactation-autoencoder-app"
version = "0.1.0"
description = "Azure Function App — TensorFlow lactation autoencoder predictions"
authors = [{ name = "Douwe de Kok", email = "douwedekok@gmail.com" }]
requires-python = ">= 3.12, < 3.13"
dependencies = [
    "azure-functions>=1.21.0",
    "fastapi>=0.115.0",
    "bovi-core",
    "lactation-autoencoder",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
